backend/requirements.txt
fastapi
uvicorn
playwright
requests
python-dotenv
openai
